my_agents/requirements.txt
openai-agents
streamlit
python-dotenv 